+++ src/main/java/com/isep/project/repository/ApplicationRepository.java
@@ -4,7 +4,7 @@ import com.isep.project.entity.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ApplicationRepository extends JpaRepository<Application, Integer>,
+public interface ApplicationRepository extends JpaRepository<Application, Long>,
         JpaSpecificationExecutor<Application>
 {
