+++ src/main/java/com/isep/project/repository/HouseRepository.java
@@ -4,7 +4,7 @@ import com.isep.project.entity.House;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface HouseRepository extends JpaRepository<House, Integer>,
+public interface HouseRepository extends JpaRepository<House, Long>,
         JpaSpecificationExecutor<House>
 {
