+++ src/main/java/com/isep/project/repository/MessageRepository.java
@@ -4,7 +4,7 @@ import com.isep.project.entity.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface MessageRepository extends JpaRepository<Message, Integer>,
+public interface MessageRepository extends JpaRepository<Message, Long>,
         JpaSpecificationExecutor<Message>
 {
